novellia_pets/requirements.txt
psycopg2-binary
django==5.0
djangorestframework
django-extensions
django-cors-headers